+++ src/main/java/me/zipestudio/tomb/mixin/ClientWorldMixin.java
@@ -21,6 +21,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.List;
+import java.util.Objects;
 
 @Mixin(ClientWorld.class)
 public class ClientWorldMixin {