+++ CHANGELOG.md
@@ -1,1 +1,5 @@
-- initial
+v1.0.1:
+- Added the `enabled` option for objects in the `tools:[ ]` block in the configuration.
+- Changed the description of the mod in the lang files.
+
+You may need to delete the `/config/tools-on-my-back.json5` file to generate a new configuration.